requirements.txt
pyspark==3.2.0
pandas>1.4.0
numpy==1.24.3
tensorflow==2.12.1
pyyaml==6.0
